Rename VAULT_BLUEPRINT to NON_FUNGIBLE_VAULT_BLUEPRINT
