+++ radix-engine-interface/src/blueprints/resource/mod.rs
@@ -2,8 +2,8 @@ mod access_rules;
 mod auth_zone;
 mod bucket;
 mod fungible;
+mod non_fungible;
 mod non_fungible_global_id;
-mod non_fungible_resource_manager;
 mod proof;
 mod proof_rule;
 mod resource;
@@ -16,8 +16,8 @@ pub use access_rules::*;
 pub use auth_zone::*;
 pub use bucket::*;
 pub use fungible::*;
+pub use non_fungible::*;
 pub use non_fungible_global_id::*;
-pub use non_fungible_resource_manager::*;
 pub use proof::*;
 pub use proof_rule::*;
 pub use resource::*;
