+++ radix-engine-interface/src/blueprints/resource/non_fungible/mod.rs
@@ -1,0 +1,5 @@
+mod non_fungible_resource_manager;
+mod non_fungible_vault;
+
+pub use non_fungible_resource_manager::*;
+pub use non_fungible_vault::*;
