+++ radix-engine-interface/src/blueprints/resource/non_fungible/non_fungible_resource_manager.rs
