+++ radix-engine-interface/src/blueprints/resource/non_fungible/non_fungible_vault.rs
@@ -1,0 +1,1 @@
+pub const NON_FUNGIBLE_VAULT_BLUEPRINT: &str = "NonFungibleVault";
