+++ radix-engine/src/blueprints/resource/vault.rs
@@ -511,7 +511,7 @@ pub struct VaultBlueprint;
 impl VaultBlueprint {
     pub fn is_vault_blueprint(blueprint: &Blueprint) -> bool {
         blueprint.package_address.eq(&RESOURCE_MANAGER_PACKAGE)
-            && (blueprint.blueprint_name.eq(VAULT_BLUEPRINT)
+            && (blueprint.blueprint_name.eq(NON_FUNGIBLE_VAULT_BLUEPRINT)
                 || blueprint.blueprint_name.eq(FUNGIBLE_VAULT_BLUEPRINT))
     }
