+++ radix-engine/src/system/client_api_impl.rs
@@ -266,7 +266,7 @@ where
                     )),
                     AllocateEntityType::Vault,
                 ),
-                VAULT_BLUEPRINT => (
+                NON_FUNGIBLE_VAULT_BLUEPRINT => (
                     RENodeInit::Object(btreemap!(
                         SubstateOffset::Vault(VaultOffset::Info) => RuntimeSubstate::VaultInfo(parser.decode_next()),
                         SubstateOffset::Vault(VaultOffset::LiquidFungible) => RuntimeSubstate::VaultLiquidFungible(parser.decode_next()),
